Remove legacy connection ID extractor
